resolve merge conflicts

+++ app/Http/Controllers/RemarksController.php
@@ -12,17 +12,20 @@ class RemarksController extends Controller
 {
     public function update(Remark $remark)
     {
-        
+        $this->authorize('update', $remark);
+
         $remark->update(request()->validate([
-            'description'=>'required|min:10|max:255|string' 
+            'description'=>'required|min:10|max:255|string'
         ]));
         return back();
     }
 
     public function destroy(Remark $remark)
     {
+        $this->authorize('delete', $remark);
+
         $remark->delete();
-        
+
         return back();
     }
 }
