+++ app/Http/Controllers/RemindersController.php
@@ -12,35 +12,15 @@ use App\LetterReminder;
 
 class RemindersController extends Controller
 {
-    public function store(Request $request)
-    {
-        $validData = request()->validate([
-            'letter_id'=>'required|exists:outgoing_letters,id',
-            'attachments' => 'required|array|max:2',
-            'attachments.*' => 'file|max:200|mimes:jpeg,jpg,png,pdf'
-        ]);
-            
-        $letter_reminder = LetterReminder::create($validData);
-
-        $letter_reminder->attachments()->createMany(
-            array_map(function ($attachedFile) {
-                return [
-                    'original_name' => $attachedFile->getClientOriginalName(),
-                    'path' => $attachedFile->store('/letter_attachments/outgoing/reminders')
-                ];
-            }, $request->file('attachments'))
-        );
-
-        return redirect()->back();
-    }
-
     public function update(LetterReminder $reminder)
     {
+        $this->authorize('update', $reminder);
+
         $data = request()->validate([
             'attachments' => 'required|array|max:2',
             'attachments.*' => 'file|max:200|mimes:jpeg,jpg,png,pdf'
         ]);
-        
+
         $reminder->attachments()->createMany(
             array_map(function ($attachedFile) {
                 return [
@@ -55,8 +35,10 @@ class RemindersController extends Controller
 
     public function destroy(LetterReminder $reminder)
     {
+        $this->authorize('delete', $reminder);
+
         $reminder->attachments->each->delete();
-        
+
         $reminder->delete();
 
         return back();
